+++ bfabric/docs/changelog.md
@@ -13,6 +13,10 @@ Versioning currently follows `X.Y.Z` where
 This release brings a new way to configure the `Bfabric` client through an environment variable completely.
 This is because it will allow round-trip passing of the configuration to subprocesses in the future.
 
+### Breaking
+
+- `default` is forbidden as an environment config name now.
+
 ### Added
 
 - New API for configuring the `Bfabric` client which **replaces** the existing `Bfabric.from_config`, which is